Move add content logic into handler
